fix route tests
