+++ alchemy/test/cloudflare/worker.test.ts
@@ -9,7 +9,6 @@ import { Self } from "../../src/cloudflare/bindings.ts";
 import { DurableObjectNamespace } from "../../src/cloudflare/durable-object-namespace.ts";
 import { KVNamespace } from "../../src/cloudflare/kv-namespace.ts";
 import { Worker, WorkerRef } from "../../src/cloudflare/worker.ts";
-import { Zone } from "../../src/cloudflare/zone.ts";
 import { destroy } from "../../src/destroy.ts";
 import { BRANCH_PREFIX } from "../util.ts";
 import { fetchAndExpectOK, fetchAndExpectStatus } from "./fetch-utils.ts";
@@ -1355,489 +1354,6 @@ describe("Worker Resource", () => {
       expect(text).toEqual("Hello, world!");
     } finally {
       await destroy(scope);
-    }
-  });
-
-  test("create worker with routes (explicit zone ID)", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-worker-routes-explicit`;
-    const testDomain = `${BRANCH_PREFIX}-routes-test.com`;
-
-    let worker: Worker | undefined;
-    let zone: Zone | undefined;
-
-    try {
-      // Create a zone first
-      zone = await Zone("test-zone", {
-        name: testDomain,
-      });
-
-      // Create a worker with explicit routes
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: `
-          export default {
-            async fetch(request, env, ctx) {
-              const url = new URL(request.url);
-              return new Response(\`Hello from worker on path: \${url.pathname}\`, {
-                status: 200,
-                headers: { 'Content-Type': 'text/plain' }
-              });
-            }
-          }
-        `,
-        format: "esm",
-        routes: [
-          {
-            pattern: `${testDomain}/api/*`,
-            zoneId: zone.id,
-          },
-          {
-            pattern: `${testDomain}/admin/*`,
-            zoneId: zone.id,
-          },
-        ],
-      });
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.routes).toBeDefined();
-      expect(worker.routes?.length).toEqual(2);
-
-      // Verify routes configuration
-      const apiRoute = worker.routes?.find((r) => r.pattern.includes("/api/"));
-      const adminRoute = worker.routes?.find((r) =>
-        r.pattern.includes("/admin/"),
-      );
-
-      expect(apiRoute).toBeDefined();
-      expect(apiRoute?.zoneId).toEqual(zone.id);
-      expect(adminRoute).toBeDefined();
-      expect(adminRoute?.zoneId).toEqual(zone.id);
-
-      // Verify routes were created via API
-      const api = await createCloudflareApi();
-      const routesResponse = await api.get(`/zones/${zone.id}/workers/routes`);
-      expect(routesResponse.ok).toBeTruthy();
-
-      const routesData: any = await routesResponse.json();
-      const createdRoutes = routesData.result.filter(
-        (route: any) => route.script === workerName,
-      );
-      expect(createdRoutes.length).toEqual(2);
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("create worker with routes (automatic zone ID inference)", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-worker-routes-auto`;
-    const testDomain = `${BRANCH_PREFIX}-auto-routes.com`;
-
-    let worker: Worker;
-    let zone: Zone;
-
-    try {
-      // Create a zone first
-      zone = await Zone("auto-zone", {
-        name: testDomain,
-      });
-
-      // Create a worker with routes that will auto-infer zone ID
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: `
-          export default {
-            async fetch(request, env, ctx) {
-              const url = new URL(request.url);
-              return new Response(\`Auto-inferred zone worker on: \${url.pathname}\`, {
-                status: 200,
-                headers: { 'Content-Type': 'text/plain' }
-              });
-            }
-          }
-        `,
-        format: "esm",
-        routes: [
-          {
-            // Zone ID should be automatically inferred from domain
-            pattern: `${testDomain}/*`,
-          },
-          {
-            // Zone ID should be inferred for subdomain too
-            pattern: `api.${testDomain}/*`,
-          },
-          {
-            // Wildcard patterns should also work
-            pattern: `*.${testDomain}/api/*`,
-          },
-        ],
-      });
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.routes).toBeDefined();
-      expect(worker.routes?.length).toEqual(3);
-
-      // Verify all routes have the inferred zone ID
-      worker.routes?.forEach((route) => {
-        expect(route.zoneId).toEqual(zone.id);
-      });
-
-      // Verify routes were created via API
-      const api = await createCloudflareApi();
-      const routesResponse = await api.get(`/zones/${zone.id}/workers/routes`);
-      expect(routesResponse.ok).toBeTruthy();
-
-      const routesData: any = await routesResponse.json();
-      const createdRoutes = routesData.result.filter(
-        (route: any) => route.script === workerName,
-      );
-      expect(createdRoutes.length).toEqual(3);
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("create worker with mixed explicit and inferred route zones", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-worker-routes-mixed`;
-    const testDomain1 = `${BRANCH_PREFIX}-mixed1.com`;
-    const testDomain2 = `${BRANCH_PREFIX}-mixed2.com`;
-
-    let worker: Worker | undefined;
-    let zone1: Zone;
-    let zone2: Zone;
-
-    try {
-      // Create two zones
-      zone1 = await Zone("mixed-zone-1", {
-        name: testDomain1,
-      });
-
-      zone2 = await Zone("mixed-zone-2", {
-        name: testDomain2,
-      });
-
-      // Create a worker with mixed explicit and auto-inferred routes
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: `
-          export default {
-            async fetch(request, env, ctx) {
-              const url = new URL(request.url);
-              return new Response(\`Mixed zones worker: \${url.hostname}\${url.pathname}\`, {
-                status: 200,
-                headers: { 'Content-Type': 'text/plain' }
-              });
-            }
-          }
-        `,
-        format: "esm",
-        routes: [
-          {
-            pattern: `${testDomain1}/*`,
-            // Zone ID will be auto-inferred
-          },
-          {
-            pattern: `${testDomain2}/*`,
-            zoneId: zone2.id, // Explicit zone ID
-          },
-          {
-            pattern: `api.${testDomain1}/*`,
-            // Zone ID will be auto-inferred from parent domain
-          },
-        ],
-      });
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.name).toEqual(workerName);
-      expect(worker.routes).toBeDefined();
-      expect(worker.routes?.length).toEqual(3);
-
-      // Verify zone IDs are correct
-      const domain1Routes = worker.routes?.filter((r) =>
-        r.pattern.includes(testDomain1),
-      );
-      const domain2Routes = worker.routes?.filter((r) =>
-        r.pattern.includes(testDomain2),
-      );
-
-      expect(domain1Routes?.length).toEqual(2);
-      expect(domain2Routes?.length).toEqual(1);
-
-      domain1Routes?.forEach((route) => {
-        expect(route.zoneId).toEqual(zone1.id);
-      });
-
-      domain2Routes?.forEach((route) => {
-        expect(route.zoneId).toEqual(zone2.id);
-      });
-
-      // Verify routes were created via API for both zones
-      const api = await createCloudflareApi();
-
-      const routes1Response = await api.get(
-        `/zones/${zone1.id}/workers/routes`,
-      );
-      const routes1Data: any = await routes1Response.json();
-      const created1Routes = routes1Data.result.filter(
-        (route: any) => route.script === workerName,
-      );
-      expect(created1Routes.length).toEqual(2);
-
-      const routes2Response = await api.get(
-        `/zones/${zone2.id}/workers/routes`,
-      );
-      const routes2Data: any = await routes2Response.json();
-      const created2Routes = routes2Data.result.filter(
-        (route: any) => route.script === workerName,
-      );
-      expect(created2Routes.length).toEqual(1);
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("fail to create worker with routes when zone cannot be inferred", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-worker-routes-fail`;
-    const nonExistentDomain = `${BRANCH_PREFIX}-nonexistent.example`;
-
-    try {
-      // Try to create worker with routes for non-existent domain
-      const workerPromise = Worker(workerName, {
-        name: workerName,
-        script: `
-          export default {
-            async fetch(request, env, ctx) {
-              return new Response('Should not be created');
-            }
-          }
-        `,
-        format: "esm",
-        routes: [
-          {
-            pattern: `${nonExistentDomain}/*`,
-            // No zoneId provided, should fail inference
-          },
-        ],
-      });
-
-      await expect(workerPromise).rejects.toThrow(
-        /Could not infer zone ID for route pattern/,
-      );
-    } finally {
-      await destroy(scope);
-      // Worker should not exist since creation failed
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("create worker with duplicate route patterns should fail", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-worker-routes-duplicate`;
-    const testDomain = `${BRANCH_PREFIX}-duplicate.com`;
-
-    let zone: Zone | undefined;
-
-    try {
-      // Create a zone
-      zone = await Zone("duplicate-zone", {
-        name: testDomain,
-      });
-
-      // Try to create worker with duplicate route patterns
-      const workerPromise = Worker(workerName, {
-        name: workerName,
-        script: `
-          export default {
-            async fetch(request, env, ctx) {
-              return new Response('Should not be created');
-            }
-          }
-        `,
-        format: "esm",
-        routes: [
-          {
-            pattern: `${testDomain}/*`,
-            zoneId: zone.id,
-          },
-          {
-            pattern: `${testDomain}/*`, // Duplicate pattern
-            zoneId: zone.id,
-          },
-        ],
-      });
-
-      await expect(workerPromise).rejects.toThrow(
-        /Duplicate route patterns found/,
-      );
-    } finally {
-      await destroy(scope);
-      // Worker should not exist since creation failed
-      await assertWorkerDoesNotExist(workerName);
-    }
-  });
-
-  test("create worker with routes that adopt existing routes", async (scope) => {
-    const workerName1 = `${BRANCH_PREFIX}-worker-routes-adopt-1`;
-    const workerName2 = `${BRANCH_PREFIX}-worker-routes-adopt-2`;
-    const testDomain = `${BRANCH_PREFIX}-adopt.com`;
-
-    let worker1: Worker | undefined;
-    let worker2: Worker | undefined;
-    let zone: Zone | undefined;
-
-    try {
-      // Create a zone
-      zone = await Zone("adopt-zone", {
-        name: testDomain,
-      });
-
-      // Create first worker with routes
-      worker1 = await Worker(workerName1, {
-        name: workerName1,
-        script: `
-          export default {
-            async fetch(request, env, ctx) {
-              return new Response('Worker 1');
-            }
-          }
-        `,
-        format: "esm",
-        routes: [
-          {
-            pattern: `${testDomain}/shared/*`,
-            zoneId: zone.id,
-          },
-        ],
-      });
-
-      expect(worker1.id).toBeTruthy();
-
-      // Create second worker that adopts the existing route
-      worker2 = await Worker(workerName2, {
-        name: workerName2,
-        script: `
-          export default {
-            async fetch(request, env, ctx) {
-              return new Response('Worker 2');
-            }
-          }
-        `,
-        format: "esm",
-        routes: [
-          {
-            pattern: `${testDomain}/shared/*`, // Same pattern
-            zoneId: zone.id,
-            adopt: true, // Should adopt existing route
-          },
-        ],
-      });
-
-      expect(worker2.id).toBeTruthy();
-      expect(worker2.routes).toBeDefined();
-      expect(worker2.routes?.length).toEqual(1);
-
-      // Verify route was adopted (script should be updated to worker2)
-      const api = await createCloudflareApi();
-      const routesResponse = await api.get(`/zones/${zone.id}/workers/routes`);
-      const routesData: any = await routesResponse.json();
-
-      const sharedRoutes = routesData.result.filter(
-        (route: any) => route.pattern === `${testDomain}/shared/*`,
-      );
-
-      // Should only be one route with this pattern
-      expect(sharedRoutes.length).toEqual(1);
-      // And it should point to worker2 now
-      expect(sharedRoutes[0].script).toEqual(workerName2);
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName1);
-      await assertWorkerDoesNotExist(workerName2);
-    }
-  });
-
-  test("update worker routes", async (scope) => {
-    const workerName = `${BRANCH_PREFIX}-worker-routes-update`;
-    const testDomain = `${BRANCH_PREFIX}-update.com`;
-
-    let worker: Worker | undefined;
-    let zone: Zone | undefined;
-
-    try {
-      // Create a zone
-      zone = await Zone("update-zone", {
-        name: testDomain,
-      });
-
-      // Create worker with initial routes
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: `
-          export default {
-            async fetch(request, env, ctx) {
-              return new Response('Initial version');
-            }
-          }
-        `,
-        format: "esm",
-        routes: [
-          {
-            pattern: `${testDomain}/api/*`,
-            zoneId: zone.id,
-          },
-        ],
-      });
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.routes?.length).toEqual(1);
-
-      // Update worker with different routes
-      worker = await Worker(workerName, {
-        name: workerName,
-        script: `
-          export default {
-            async fetch(request, env, ctx) {
-              return new Response('Updated version');
-            }
-          }
-        `,
-        format: "esm",
-        routes: [
-          {
-            pattern: `${testDomain}/api/v2/*`, // Different pattern
-            zoneId: zone.id,
-          },
-          {
-            pattern: `${testDomain}/admin/*`, // Additional route
-            zoneId: zone.id,
-          },
-        ],
-      });
-
-      expect(worker.id).toBeTruthy();
-      expect(worker.routes?.length).toEqual(2);
-
-      // Verify new routes were created
-      const api = await createCloudflareApi();
-      const routesResponse = await api.get(`/zones/${zone.id}/workers/routes`);
-      const routesData: any = await routesResponse.json();
-
-      const workerRoutes = routesData.result.filter(
-        (route: any) => route.script === workerName,
-      );
-
-      expect(workerRoutes.length).toEqual(2);
-
-      const patterns = workerRoutes.map((route: any) => route.pattern);
-      expect(patterns).toContain(`${testDomain}/api/v2/*`);
-      expect(patterns).toContain(`${testDomain}/admin/*`);
-      expect(patterns).not.toContain(`${testDomain}/api/*`); // Old route should be removed
-    } finally {
-      await destroy(scope);
-      await assertWorkerDoesNotExist(workerName);
     }
   });
 });
